Add analystics backend to the agent
There's one simple "start" event now that ships the main version.

On every new event, we'll include the origin and session_id so we can
track events. A lot of this is still new, so i'm not sure which events
we'll want to gather and how we'll do analytics and what not.